test(range): migrate legacy basic
