+++ core/src/components/range/test/legacy/basic/range.e2e.ts-snapshots/range-diff-ios-ltr-Mobile-Safari-linux.png
