+++ core/src/components/range/test/legacy/basic/range.e2e.ts-snapshots/range-diff-md-rtl-Mobile-Firefox-linux.png
